+++ CHANGELOG.md
@@ -4,15 +4,34 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.dev14]
 
+
 ### Added
 
 * Add snappy decompression for parquet; `pyarrow` instead of `fastparquet` (#193)
 * Password-less sudo for panoptes user on dev docker image (#193)
 * `get_metadata` has an optional progress bar. (#194)
+* Add `bayer.get_stamp_slice` for getting a stamp slice while respecting the superpixel. This was removed awhile ago and has been re-added and improved. (#196)
+  * Adjusting the offsets so the center pixel is always:
+    
+    ```
+       G2 B
+       R  G1
+    ```
 
 ### Bug fixes
 
 * Fix time-based search (#193)
+
+### Changed
+
+* **Breaking** Only support getting stars directly from the WCS, not the footprint. (#194)
+  * `get_stars_from_footprint` -> `get_stars_from_wcs`
+  * Better logging
+  * Consistent column names for dtypes
+  * Vmag bin comes from sql.
+  * Allow for different RA/Dec column names.
+  * Better catalog match function.
+* `sextractor` param changes. (#194)
 
 ## [0.2.13] - 2020-05-14